Add safeguard against direct calls to setText method
Introduced a flag, updatingText, to prevent unnecessary actions when the setText() call is not made from the setRawPhoneNumber() function. Updated textProperty().addListener to check if setText() is called indirectly or directly to avoid overwriting the raw phone number. The flag manages this condition, it's set to true before making changes to avoid reverting them. It's reset back to false in the finally block after changes are made, ready for the next call.